requirements.txt
aiogram~=3.5
aiosqlite~=0.19
apscheduler~=3.10
httpx[http2]~=0.27
pydantic~=2.7
python-dotenv~=1.0
